refactor(middleware): move built-ins into builtin/, split security 7 ways
Concrete middleware move out of middleware.py and middleware_ext/ into
builtin/, which depends on core/ only -- never on stack/ -- so the stack can
import built-ins without a cycle.

security.py was 1,185 lines holding seven unrelated controls. CORS and CSRF
change for different reasons and get reviewed by different people:

  security/{cors,csp,csrf,hsts,https_redirect,proxy_fix,headers}.py

ExceptionMiddleware drops from 270 lines to ~200 by moving content
negotiation and the fault-to-status table to utils/, and by splitting one
150-line except-block into a handler per exception family.

middleware_ext/ stays as a re-export shim that warns once per process.
User workspace.py files name these as dotted strings --
".use('aquilia.middleware_ext.SecurityHeadersMiddleware')" -- and those are
resolved by importlib at boot, so deleting the package would fail other
people's projects with an ImportError rather than a deprecation path.

Removed as dead weight:

- OptionalSessionMiddleware: allocated a fresh SessionMiddleware per request
  purely to delegate to it. SessionMiddleware now takes session_engine=None
  and uses the new should_run hook, which is the same behaviour without the
  per-request allocation or the second class.
- create_session_middleware: a factory for one product, selected by a bool
  flag. Zero callers.
