+++ aquilia/middleware/builtin/static.py
@@ -28,7 +28,7 @@ from typing import (
     TYPE_CHECKING,
 )
 
-from aquilia.middleware import Middleware
+from aquilia.middleware.core.base import Middleware
 
 if TYPE_CHECKING:
     from aquilia.controller.base import RequestCtx
